+++ iblutil/io/net/base.py
@@ -9,9 +9,17 @@ from abc import ABC, abstractmethod
 from urllib.parse import urlparse
 import urllib.request
 import ipaddress
-from enum import IntEnum
+from operator import or_
+from functools import reduce
+from enum import IntFlag, IntEnum, auto  # py3.11 STRICT
 
-LISTEN_PORT = 10001  # listen for commands on this port
+LISTEN_PORT = 11001  # listen for commands on this port
+PROTOCOL_VERSION = '1.0.0'  # Versioning for ExpMessage, ExpStatus enumerations, and Service base class
+
+
+def is_success(future: asyncio.Future) -> bool:
+    """Check if future successfully resolved."""
+    return future.done() and not future.cancelled() and future.exception() is None
 
 
 def external_ip():
@@ -137,36 +145,48 @@ def validate_uri(uri, resolve_host=True, default_port=LISTEN_PORT, default_proc=
     return f'{proc or default_proc}://{host}:{port}'
 
 
-class ExpMessage(IntEnum):
+# class ExpMessage(IntFlag, boundary=STRICT):  # py3.11
+class ExpMessage(IntFlag):
     """A set of standard experiment messages for communicating between rigs."""
 
     """Experiment is initializing."""
-    EXPINIT = 10
+    EXPINIT = auto()
     """Experiment has begun."""
-    EXPSTART = 20
+    EXPSTART = auto()
     """Experiment has stopped."""
-    EXPEND = 30
+    EXPEND = auto()
     """Experiment cleanup begun."""
-    EXPCLEANUP = 40
+    EXPCLEANUP = auto()
     """Experiment interrupted."""
-    EXPINTERRUPT = 50
+    EXPINTERRUPT = auto()
     """Experiment status."""
-    EXPSTATUS = 1
+    EXPSTATUS = auto()
     """Experiment info, including task protocol start and end."""
-    EXPINFO = 2
+    EXPINFO = auto()
     """Alyx token."""
-    ALYX = 3
+    ALYX = auto()
+    __version__ = PROTOCOL_VERSION
+
+    @classmethod
+    def any(cls) -> 'ExpMessage':
+        """Return enumeration comprising all possible messages.
+
+        NB: This doesn't include the null ExpMessage (0), used to indicate API errors.
+        """
+        return reduce(or_, cls)
 
     @staticmethod
-    def validate(event):
+    def validate(event, allow_bitwise=True):
         """
-        Validate an event message, returning an corresponding enumeration if valid and raising an
+        Validate an event message, returning a corresponding enumeration if valid and raising an
         exception if not.
 
         Parameters
         ----------
         event : str, int, ExpMessage
             An event message to validate.
+        allow_bitwise : bool
+            If false, raise if event is the result of a bitwise operation.
 
         Returns
         -------
@@ -179,7 +199,7 @@ class ExpMessage(IntEnum):
             event is neither a string, integer nor enumeration.
         ValueError
             event does not correspond to any ExpMessage enumeration, neither in its integer form
-            nor its string name.
+            nor its string name, or `allow_bitwise` is false and value is combination of events.
 
         Examples
         --------
@@ -202,12 +222,41 @@ class ExpMessage(IntEnum):
                     raise TypeError(f'Unknown event type {type(event)}')
             except KeyError:
                 raise ValueError(f'Unrecognized event "{event}". '
-                                 f'Choices: {tuple(map(lambda item: item.name, ExpMessage))}')
+                                 f'Choices: {tuple(ExpMessage.__members__.keys())}')
+        if not allow_bitwise and event not in list(ExpMessage):
+            raise ValueError('Compound (bitwise) events not permitted. '
+                             f'Choices: {tuple(ExpMessage)}')
         return event
+
+    def __iter__(self):  # py3.11 remove this method
+        """Iterate over the individual bits in the enumeration.
+
+        NB: This method is copied from Python 3.11 which supports iteration of Enum objects.
+        """
+        num = self.value
+        while num:
+            b = num & (~num + 1)
+            yield b
+            num ^= b
+
+
+class ExpStatus(IntEnum):
+    """A set of standard statuses for communicating between rigs."""
+
+    """Service is connected."""
+    CONNECTED = 0
+    """Service is initialized."""
+    INITIALIZED = 10
+    """Service is running."""
+    RUNNING = 20
+    """Experiment has stopped."""
+    STOPPED = 30
+    __version__ = PROTOCOL_VERSION
 
 
 class Service(ABC):
     """An abstract base class for auxiliary experiment services."""
+    __version__ = PROTOCOL_VERSION
     __slots__ = 'name'
 
     @abstractmethod
@@ -256,6 +305,9 @@ class Service(ABC):
         any, None
             Optional extra data.
         """
+        exp_ref = exp_ref or None
+        if isinstance(exp_ref, dict):
+            exp_ref = '_'.join(map(str, (exp_ref['date'], int(exp_ref['sequence']), exp_ref['subject'])))
         return ExpMessage.EXPSTART, exp_ref, data
 
     @abstractmethod
@@ -281,6 +333,59 @@ class Service(ABC):
             Optional extra data.
         """
         return ExpMessage.EXPINTERRUPT if immediately else ExpMessage.EXPEND, data
+
+    @abstractmethod
+    def status(self, status):
+        """
+        Report experiment status.
+
+        NB: This is intended to be lightweight. For more detail and custom data use the info
+        method.
+
+        This is intended to specify the expected message signature. The subclassed method should
+        serialize the returned values and pass them to the transport layer.
+
+        Parameters
+        ----------
+        status : ExpStatus
+            The experiment status enumeration.
+
+        Returns
+        -------
+        ExpMessage.EXPSTATUS
+            The EXPSTATUS event.
+        ExpStatus
+            The validated experiment status.
+        """
+        if not isinstance(status, ExpStatus):
+            status = ExpStatus(status) if isinstance(status, int) else ExpStatus[status]
+        return ExpMessage.EXPSTATUS, status
+
+    @abstractmethod
+    def info(self, status, data=None):
+        """
+        Report experiment information.
+
+        This is intended to specify the expected message signature. The subclassed method should
+        serialize the returned values and pass them to the transport layer.
+
+        Parameters
+        ----------
+        status : ExpStatus
+            The experiment status enumeration.
+        data : any
+            Optional extra data to send to the remote server.
+
+        Returns
+        -------
+        ExpMessage.EXPINFO
+            The EXPINFO event.
+        ExpStatus
+            The validated experiment status.
+        any, None
+            Optional extra data.
+        """
+        return ExpMessage.EXPINFO, Service.status(self, status)[1], data
 
     @abstractmethod
     def cleanup(self, data=None):
@@ -326,7 +431,9 @@ class Service(ABC):
         dict
             The Alyx token in the form {user: token}.
         """
-        return ExpMessage.ALYX, alyx.base_url, alyx._par.TOKEN
+        base_url = alyx.base_url if alyx else None
+        token = {alyx.user: alyx._token} if alyx and alyx.is_logged_in else {}
+        return ExpMessage.ALYX, base_url, token
 
 
 class Communicator(Service):
@@ -339,13 +446,14 @@ class Communicator(Service):
     server_uri : str
         The full URI of the remote device, e.g. udp://192.168.0.1:1001
     """
-    __slots__ = ('server_uri', '_callbacks', 'logger')
+    __slots__ = ('server_uri', '_callbacks', 'logger', 'name')
 
     def __init__(self, server_uri, name=None, logger=None):
         self.server_uri = validate_uri(server_uri)
         self.name = name or server_uri
         self.logger = logger or logging.getLogger(self.name)
-        self._callbacks = dict(map(lambda item: (item, []), ExpMessage))
+        # Init callbacks map of ExpMessage -> list, including null ExpMessage for processing callback errors
+        self._callbacks = dict(map(lambda item: (item, []), (ExpMessage(0), *ExpMessage.__members__.values())))
 
     def assign_callback(self, event, callback):
         """
@@ -369,9 +477,14 @@ class Communicator(Service):
         event = ExpMessage.validate(event)
         if not (callable(callback) or isfuture(callback)):
             raise TypeError('Callback must be callable or a Future')
-        self._callbacks.setdefault(event, []).append(callback)
+        if event is ExpMessage(0):
+            self._callbacks.setdefault(event, []).append((callback, False))
+        else:
+            return_event = event not in ExpMessage
+            for e in event:  # iterate over enum as bitwise ops permitted
+                self._callbacks.setdefault(e, []).append((callback, return_event))
 
-    def clear_callbacks(self, event, callback=None):
+    def clear_callbacks(self, event, callback=None, cancel_futures=True):
         """
         For a given event, remove the provided callback, or all callbacks if none were provided.
 
@@ -381,6 +494,8 @@ class Communicator(Service):
             The event for which the callback was registered.
         callback : function, asyncio.Future
             The callback or future to remove.
+        cancel_futures : bool
+            If True and callback is a Future, cancel before removing.
 
         Returns
         -------
@@ -391,19 +506,26 @@ class Communicator(Service):
         event = ExpMessage.validate(event)
         if callback:  # clear specific callback
             # Wrapped callables have an id attribute containing the hash of the inner function
-            ids = [getattr(cb, 'id', None) or hash(cb) for cb in self._callbacks[event]]
-            cb_id = getattr(callback, 'id', None) or hash(callback)
-            while True:
-                try:
-                    idx = ids.index(cb_id)
-                    self._callbacks[event].pop(idx)
-                    ids.pop(idx)
-                    i += 1
-                except (IndexError, ValueError):
-                    break
+            for evt in event:  # iterate as bitwise enums permitted, e.g. ~ExpMessage.ALYX
+                ids = [getattr(cb, 'id', None) or hash(cb) for cb, _ in self._callbacks[evt]]
+                cb_id = getattr(callback, 'id', None) or hash(callback)
+                while True:
+                    try:
+                        idx = ids.index(cb_id)
+                        if cancel_futures and isfuture(cb := self._callbacks[evt][idx][0]):
+                            cb.cancel()
+                        del self._callbacks[evt][idx]
+                        del ids[idx]
+                        i += 1
+                    except (IndexError, ValueError):
+                        break
         else:  # clear all callbacks for event
-            i = len(self._callbacks[event])
-            self._callbacks[event] = []
+            for evt in event:
+                if cancel_futures:
+                    for cb in filter(isfuture, map(lambda x: x[0], self._callbacks[evt])):
+                        cb.cancel()
+                i += len(self._callbacks[evt])
+                del self._callbacks[evt][:]
         self.logger.debug('[%s] %i callbacks cleared', self.name, i)
         return i
 
@@ -423,11 +545,15 @@ class Communicator(Service):
 
         Examples
         --------
-        >>> data = await com.on_event('EXPSTART')
+        >>> data, addr = await com.on_event('EXPSTART')
 
         >>> event = await asyncio.create_task(com.on_event('EXPSTART'))
         >>> ...
         >>> data = await event
+
+        Await more than one event
+
+        >>> data, addr, event = await com.on_event(ExpMessage.EXPEND | ExpMessage.EXPINTERRUPT)
         """
         loop = asyncio.get_running_loop()
         fut = loop.create_future()
@@ -437,7 +563,7 @@ class Communicator(Service):
     @property
     def port(self) -> int:
         """int: the remote port"""
-        return urlparse(self.server_uri).port
+        return int(urlparse(self.server_uri).port)
 
     @property
     def hostname(self) -> str:
@@ -485,18 +611,26 @@ class Communicator(Service):
         data = self.decode(data)
         if isinstance(data, (list, tuple)) and len(data) > 1:
             event, *data = data
-            for f in self._callbacks[ExpMessage.validate(event)].copy():
+            event = ExpMessage.validate(event, allow_bitwise=False) if event else ExpMessage(0)
+            if event is ExpMessage(0):
+                # An error in the remote callback function occurred
+                err, evt = data
+                self.logger.error('Callback for %s on %s://%s:%i failed with %s',
+                                  ExpMessage(evt).name, self.protocol, *addr, err)
+            for f, return_event in self._callbacks[event].copy():
                 if isfuture(f):
-                    if not f.cancelled():
-                        f.set_result((data, addr))
+                    if f.done():
+                        self.logger.warning('Future %s already resolved', f)
+                    elif not f.cancelled():
+                        f.set_result((data, addr, event) if return_event else (data, addr))
                     self.clear_callbacks(event, f)  # Remove future from list
                 else:
                     try:
-                        f(data, addr)
+                        f(data, addr, event) if return_event else f(data, addr)
                     except Exception as ex:
                         self.logger.error('Callback "%s" failed with error "%s"', f, ex)
-                        message = self.encode([ExpMessage.EXPINTERRUPT, f'{type(ex).__name__}: {ex}'])
-                        self.send(message, addr)  # FIXME should be confirmed send
+                        message = self.encode([0, f'{type(ex).__name__}: {ex}', event])
+                        self.send(message, addr)
                         break
         else:
             warnings.warn(f'Expected list, got {data}', RuntimeWarning)
@@ -552,6 +686,6 @@ class Communicator(Service):
     def close(self):
         """De-register all callbacks and cancel futures"""
         for event in self._callbacks:
-            for fut in filter(isfuture, self._callbacks[event]):
+            for fut in filter(isfuture, map(lambda x: x[0], self._callbacks[event])):
                 fut.cancel('Close called on communicator')
-            self._callbacks[event] = []
+            del self._callbacks[event][:]