+++ tests/test_net.py
@@ -1,17 +1,19 @@
+import sys
 import asyncio
 import logging
 import unittest
 from unittest import mock
 import ipaddress
 import socket
+from packaging.version import Version
 
 from iblutil.io.net import base, app
 
-import sys
-ver = float('%i.%i' % sys.version_info[:2])
+ver = (getattr(sys.version_info, v) for v in ('major', 'minor', 'micro'))
+ver = Version('.'.join(map(str, ver)))
 
 
-class TestBase(unittest.TestCase):
+class TestBase(unittest.IsolatedAsyncioTestCase):
     """Test for base network utils.
 
     NB: This requires internet access.
@@ -26,7 +28,7 @@ class TestBase(unittest.TestCase):
         self.assertEqual(expected, base.validate_uri(uri[:-5], default_port=9999))
         uri = base.validate_uri(ipaddress.ip_address('192.168.0.1'), default_port=9999)
         self.assertEqual(expected, uri)
-        self.assertEqual('udp://foobar:10001', base.validate_uri('foobar', resolve_host=False))
+        self.assertEqual('udp://foobar:11001', base.validate_uri('foobar', resolve_host=False))
         # Check IP resolved
         uri = base.validate_uri('http://google.com:80', resolve_host=True)
         expected = (ipaddress.IPv4Address, ipaddress.IPv6Address)
@@ -47,13 +49,13 @@ class TestBase(unittest.TestCase):
         self.assertFalse(ipaddress.ip_address(base.external_ip()).is_private)
 
     def test_ExpMessage(self):
-        """Test for ExpMessage.validate method"""
+        """Test for ExpMessage.validate method."""
         # Check identity
         msg = base.ExpMessage.validate(base.ExpMessage.EXPINFO)
         self.assertIs(msg, base.ExpMessage.EXPINFO)
 
         # Check integer input
-        msg = base.ExpMessage.validate(40)
+        msg = base.ExpMessage.validate(int(base.ExpMessage.EXPCLEANUP))
         self.assertIs(msg, base.ExpMessage.EXPCLEANUP)
 
         # Check string input
@@ -65,6 +67,12 @@ class TestBase(unittest.TestCase):
             base.ExpMessage.validate(b'EXPSTART')
         with self.assertRaises(ValueError):
             base.ExpMessage.validate('EXPSTOP')
+
+        # Test allow_bitwise kwarg
+        event = base.ExpMessage.any()
+        self.assertIs(base.ExpMessage.validate(event), event)
+        with self.assertRaises(ValueError):
+            base.ExpMessage.validate(event, allow_bitwise=False)
 
     def test_encode(self):
         """Tests for iblutil.io.net.base.Communicator.encode"""
@@ -82,8 +90,22 @@ class TestBase(unittest.TestCase):
             decoded = base.Communicator.decode(data + b'"')
             self.assertEqual(decoded, '[null, 21, "message"]"')
 
+    async def test_is_success(self):
+        """Tests for iblutil.io.net.base.is_success function."""
+        # Expect True only when set_result has been called.
+        fut = asyncio.get_event_loop().create_future()
+        self.assertFalse(base.is_success(fut))
+        fut.set_result(None)
+        self.assertTrue(base.is_success(fut))
+        fut = asyncio.get_event_loop().create_future()
+        fut.cancel()
+        self.assertFalse(base.is_success(fut))
+        fut = asyncio.get_event_loop().create_future()
+        fut.set_exception(RuntimeError)
+        self.assertFalse(base.is_success(fut))
 
-@unittest.skipIf(ver < 3.9, 'only version 3.9 or later supported')
+
+@unittest.skipIf(ver < Version('3.9'), 'only version 3.9 or later supported')
 class TestUDP(unittest.IsolatedAsyncioTestCase):
 
     async def asyncSetUp(self):
@@ -91,7 +113,7 @@ class TestUDP(unittest.IsolatedAsyncioTestCase):
         self.client = await app.EchoProtocol.client('localhost', name='client')
 
     async def test_start(self):
-        """Tests confirmed send via start command"""
+        """Tests confirmed send via start command."""
         # Check socket type
         self.assertIs(self.server._socket.type, app.socket.SOCK_DGRAM)
         self.assertIs(self.client._socket.type, app.socket.SOCK_DGRAM)
@@ -101,27 +123,31 @@ class TestUDP(unittest.IsolatedAsyncioTestCase):
         exp_ref = '2022-01-01_1_subject'
         with self.assertLogs(self.server.logger, logging.INFO) as log:
             await self.client.start(exp_ref)
-            self.assertIn(f'Received \'[20, "{exp_ref}", null]', log.records[-1].message)
+            self.assertIn(f'Received \'[{base.ExpMessage.EXPSTART}, "{exp_ref}", null]', log.records[-1].message)
         spy.assert_called_with([exp_ref, None], (self.client._socket.getsockname()))
 
     async def test_callback_error(self):
-        """Tests behaviour when callback raises exception"""
+        """Tests behaviour when callback raises exception."""
         callback = mock.MagicMock()
         callback.side_effect = ValueError('Callback failed')
 
         self.server.assign_callback('expinit', callback)
-        task = asyncio.create_task(self.server.on_event('expinterrupt'))
+        task = asyncio.create_task(self.client.on_event(0))
         with self.assertLogs(self.server.logger, logging.ERROR) as log:
-            base.Communicator._receive(self.server, b'[10, null]', self.client._socket.getsockname())
+            base.Communicator._receive(self.server, b'[1, null]', self.client._socket.getsockname())
             self.assertEqual(1, len(log.records))
             self.assertIn('Callback failed', log.records[-1].message)
 
         # Check error propagated back to client
-        (err,), _ = await task
+        with self.assertLogs(self.client.logger, logging.ERROR) as log:
+            (err, evt), _ = await task
+            self.assertEqual(1, len(log.records))
+            self.assertIn('Callback failed', log.records[-1].message)
+        self.assertEqual(base.ExpMessage.EXPINIT.value, evt)
         self.assertIn('Callback failed', err)
 
     async def test_on_event(self):
-        """Test on_event method as well as init, stop and cleanup"""
+        """Test on_event method as well as init, stop, etc."""
         # INIT
         task = asyncio.create_task(self.server.on_event('expinit'))
         await self.client.init(42)
@@ -146,9 +172,101 @@ class TestUDP(unittest.IsolatedAsyncioTestCase):
         actual, _ = await task
         self.assertEqual(['foo'], actual)
 
+        # START
+        task = asyncio.create_task(self.server.on_event('expstart'))
+        await self.client.start('2020-01-01_1_baz', {'foo': 'bar'})
+        actual, _ = await task
+        self.assertEqual(['2020-01-01_1_baz', {'foo': 'bar'}], actual)
+
+        # STATUS
+        task = asyncio.create_task(self.server.on_event('EXPSTATUS'))
+        await self.client.status(base.ExpStatus.STOPPED)
+        actual, _ = await task
+        self.assertEqual([base.ExpStatus.STOPPED.value], actual)
+
+        # INFO
+        task = asyncio.create_task(self.server.on_event('expinfo'))
+        await self.client.info(base.ExpStatus.RUNNING, {'foo': 'bar'})
+        actual, _ = await task
+        self.assertEqual([base.ExpStatus.RUNNING.value, {'foo': 'bar'}], actual)
+
+    async def test_alyx(self):
+        """Test iblutil.io.net.app.EchoProtocol.alyx method."""
+        # Mock an AlyxClient instance that is logged in
+        alyx = mock.MagicMock()
+        alyx.is_logged_in = True
+        alyx.base_url = 'https://alyx.website.net'
+        alyx.user = 'foo'
+        alyx._token = {'token': '4157aa522b855239cd05f4d23d40563aa0518359'}
+
+        # When Alyx instance is logged in, expect the token to be broadcast
+        task = asyncio.create_task(self.server.on_event('ALYX'))
+        self.assertIsNone(await self.client.alyx(alyx), 'unexpected argument returned')
+        actual, _ = await task  # wait for server process request
+        expected = ['https://alyx.website.net', {'foo': {'token': '4157aa522b855239cd05f4d23d40563aa0518359'}}]
+        self.assertEqual(expected, actual, 'client failed to send alyx token to server')
+
+        # Client should request and return a token when Alyx instance is not logged in
+        async def _req_callback():
+            data, addr = await self.server.on_event('ALYX')
+            self.assertEqual(['https://alyx.website.net', {}], data)
+            await self.server.alyx(alyx, addr)
+
+        # Mock an AlyxClient instance that is not logged in
+        alyx_logged_out = mock.MagicMock()
+        alyx_logged_out.is_logged_in = False
+        alyx_logged_out.base_url = 'https://alyx.website.net'
+        alyx_logged_out.user = alyx_logged_out._token = None
+
+        task = asyncio.create_task(_req_callback())
+        token = await self.client.alyx(alyx_logged_out)
+        await task  # wait for server process request
+        self.assertEqual(expected, token, 'failed to return requested token from server')
+
+        # Try the same thing but with None instead of an Alyx instance
+        async def _req_callback():
+            data, addr = await self.server.on_event('ALYX')
+            self.assertEqual([None, {}], data)
+            await self.server.alyx(alyx, addr)
+
+        task = asyncio.create_task(_req_callback())
+        token = await self.client.alyx(None)
+        await task  # wait for server process request
+        self.assertEqual(expected, token)
+
+    async def test_confirmed_send_validation(self):
+        """Basic tests for iblutil.io.net.app.EchoProtocol.confirmed_send exception handling."""
+        # Expect to raise in server role when no address provided
+        with self.assertRaises(TypeError):
+            await self.server.confirmed_send(None)
+        # Expect to raise in client role when provided address does not match server URI
+        with self.assertRaises(ValueError):
+            await self.client.confirmed_send(None, addr=('localhost', self.client.port + 100))
+        # Expect to raise when echo timeout is 0
+        try:
+            self.client.default_echo_timeout = 0
+            with self.assertRaises(ValueError):
+                await self.client.confirmed_send(None)
+        finally:
+            self.client.default_echo_timeout = app.EchoProtocol.default_echo_timeout
+        # Expect timeout arg to override default echo timeout, expect error raised on timeout
+        assert self.client.is_connected
+        with mock.patch('iblutil.io.net.app.asyncio.wait_for', side_effect=asyncio.TimeoutError) as m, \
+                self.assertRaises(TimeoutError):
+            await self.client.confirmed_send(None, timeout=0.2)
+            self.assertFalse(self.client.is_connected, 'failed to close communicator on echo timeout error')
+        m.assert_awaited_once_with(mock.ANY, timeout=0.2)
+        # Expect to raise RuntimeError with explanation when messages don't match
+        with mock.patch('iblutil.io.net.app.asyncio.wait_for', side_effect=RuntimeError), \
+                self.assertRaises(RuntimeError) as cm:
+            await self.client.confirmed_send(None)
+        self.assertIn('unexpected response', str(cm.exception).lower())
+
     def test_communicator(self):
-        """Basic tests for iblutil.io.net.app.EchoProtocol"""
+        """Basic tests for iblutil.io.net.app.EchoProtocol, namely the role setter."""
         # Check role validation
+        self.assertEqual(self.server.role, 'server')
+        self.assertEqual(self.client.role, 'client')
         with self.assertRaises(ValueError):
             app.EchoProtocol('localhost', 'foo')
         with self.assertRaises(AttributeError):
@@ -215,7 +333,7 @@ class TestUDP(unittest.IsolatedAsyncioTestCase):
         self.server.close()
 
 
-@unittest.skipIf(ver < 3.9, 'only version 3.9 or later supported')
+@unittest.skipIf(ver < Version('3.9'), 'only version 3.9 or later supported')
 class TestWebSockets(unittest.IsolatedAsyncioTestCase):
     """Test net.app.EchoProtocol with a TCP/IP transport layer"""
 
@@ -224,7 +342,7 @@ class TestWebSockets(unittest.IsolatedAsyncioTestCase):
         self.client = await app.EchoProtocol.client('ws://localhost:8888', name='client')
 
     async def test_start(self):
-        """Tests confirmed send via start command"""
+        """Tests confirmed send via start command."""
         # Check socket indeed TCP
         self.assertIs(self.server._socket.type, app.socket.SOCK_STREAM)
         self.assertIs(self.client._socket.type, app.socket.SOCK_STREAM)
@@ -235,7 +353,7 @@ class TestWebSockets(unittest.IsolatedAsyncioTestCase):
         exp_ref = '2022-01-01_1_subject'
         with self.assertLogs(self.server.logger, logging.INFO) as log:
             await self.client.start(exp_ref)
-            self.assertIn(f'Received \'[20, "{exp_ref}", null]', log.records[-1].message)
+            self.assertIn(f'Received \'[{base.ExpMessage.EXPSTART}, "{exp_ref}", null]', log.records[-1].message)
         spy.assert_called_with([exp_ref, None], (self.client._socket.getsockname()))
 
     def test_send_validation(self):
@@ -262,7 +380,7 @@ class TestWebSockets(unittest.IsolatedAsyncioTestCase):
         self.server.close()
 
 
-@unittest.skipIf(ver < 3.9, 'only version 3.9 or later supported')
+@unittest.skipIf(ver < Version('3.9'), 'only version 3.9 or later supported')
 class TestServices(unittest.IsolatedAsyncioTestCase):
     """Tests for the app.Services class"""
 
@@ -307,7 +425,7 @@ class TestServices(unittest.IsolatedAsyncioTestCase):
             await self.server_1.init('foo', addr=addr)
 
         self.assertEqual(2, callback.call_count)
-        callback.assert_called_with(['foo'], ('127.0.0.1', 10001))
+        callback.assert_called_with(['foo'], ('127.0.0.1', 11001))
 
         # Check return_service arg
         callback2 = mock.MagicMock(spec_set=True)
@@ -315,7 +433,7 @@ class TestServices(unittest.IsolatedAsyncioTestCase):
         for addr in map(lambda x: x._socket.getsockname(), clients):
             await self.server_1.init('foo', addr=addr)
         self.assertEqual(2, callback2.call_count)
-        callback2.assert_called_with(['foo'], ('127.0.0.1', 10001), self.client_2)
+        callback2.assert_called_with(['foo'], ('127.0.0.1', 11001), self.client_2)
 
         # Check validation
         with self.assertRaises(TypeError):